Secure Storage: Support extending file size and file hole
With this patch, user can extend file size by truncate operation.

Also the file hole feature is supported in this patch, that is,
user can seek to a position larger than file size and write data
from the position.

The file hole implementation actually is not an efficient solution,
because we always fill zero in these holes.

Tested-by: James Kung <james.kung@linaro.org> (QEMU, MT8173 EVB)
Tested-by: Pascal Brand <pascal.brand@linaro.org> (QEMU platform)
Tested-by: Jerome Forissier <jerome.forissier@linaro.org> (HiKey)

+++ core/include/tee/tee_fs.h
@@ -58,7 +58,7 @@
  */
 #define REE_FS_NAME_MAX (TEE_FS_NAME_MAX + 20)
 
-typedef int tee_fs_off_t;
+typedef int64_t tee_fs_off_t;
 typedef uint32_t tee_fs_mode_t;
 typedef struct tee_fs_dir tee_fs_dir;
